requirements.txt
dvc
pyyaml
beautifulsoup4
requests
pandas